Implement Dir::create_dir
For the time being, only targeting *nix and WASI. Underneath, we're
using `yanix::mkdirat` which translates to either the POSIX syscall
in *nixes, or to `path_create_directory` if linked against WASI's libc.

While here, I've added `DirEntry::path` method to make the example
compile and run. I'm not sure what your thoughts are on this, but
IMHO, to be compatible-ish with Rust's `libstd`, we could make `path`
equivalent to `file_name` but with `./` prepended to the final result.
Alternatively, we could scrap the prepending altogether and have `path`
simply return `OsString` returned by `file_name` wrapped in `PathBuf`
(although this will cost us an additional alloc).