Refactor check for openat2 on Linux
This commit swaps the use of `lazy_static` to store the pointer to
a function (which can either be `openat2` or `open_manully_wrapper`),
re-use Alex's idea to store a predicate in `AtomicBool` which is set
to `false` initially. Upon calling `openat2` for the first time, if
the host returns `ENOSYS`, swap the bool for `false` and from that
point on, always fall back to `open_manually_wrapper`.