Refactor openat2 (#13)
* Refactor check for openat2 on Linux

This commit swaps the use of `lazy_static` to store the pointer to
a function (which can either be `openat2` or `open_manully_wrapper`),
re-use Alex's idea to store a predicate in `AtomicBool` which is set
to `false` initially. Upon calling `openat2` for the first time, if
the host returns `ENOSYS`, swap the bool for `false` and from that
point on, always fall back to `open_manually_wrapper`.

* Fix debug assert in openat2 call

Also, add host-dependent `Dir::open_file` since we need to be able
to tweak mode bits on Linux when using `openat2` syscall. Otherwise,
we'll run into `EINVAL` errnos, since `openat2` is very strict about
passing incorrect `open_how` structure.

* Unify error msgs for escape attempt

* Check for empty path before calling openat2

* If EXDEV, check if path abs, and remap error if so

* Add srht archlinux CI script

* Add srht badge to README

* Swapped svg with link to CI around

* Address comments

* Fix formatting and remove obsolete comments

* Fix tests

* Fix tests (again)