Remove some TODO comments (#17)
* Resolve a TODO question.

`create_dir` shouldn't return a `Dir`. It'd be nice if it could, but
that's a deviation from `libstd` and from what underlying OS's actually
do. If someone wants to create a directory and not open it, this would
add overhead for them.

* Remove AT_EMPTY_PATH TODO comments.

AT_EMPTY_PATH only works on Linux and only >= 2.6.39, and isn't that
valuable, so don't bother.

* Drop the question about implementing Unix APIs on Windows for now.

The goal is to mirror std, not provide extra features.