Use `open` in `open_parent`.
Use regular `open` in `open_parent`, to take advantage of optimizations
such as `openat2`.

This also adjusts the interface of `open_manually` and `open_parent` to
operate in terms of `MaybeOwnedFile` arguments and return types rather
than mutable references, which makes working with the borrow checker
a little easier.

`stat` still needs `open_manually`, so that it can use a a single symlink
count across multple calls, so create a separate `open_parent_manually`
for it.