Adjust the access check for `..` to avoid sending `..` to the OS.
When checking whether we have permissions to search for `..` within a
directory, instead test for `.`. The results should always be the same,
and this avoids sending a `..` path to the host OS. This avoids
triggering an `ENOTCAPABLE` when running under capability-mode in
Capsicum, and is generally cleaner since `..` paths often require
special considerations.

Fixes #122.