Add a `MetadataExt` feature.
This renames the `cap-dir-ext` crate to `cap-fs-ext`, and adds a
`MetadataExt` extension trait to it, exposing methods to return
`dev()`, `ino()`, and `nlink()` in a portable way.

This requires some reorganization because the needed features are
available in nightly Rust, where we want to use them, but we need
to support stable Rust as well, so we use the `winx` crate to do
the work ourselves on stable.