Expose cap_std::fs_utf8 in docs.rs
I was trying to use `fs_utf8` for some of my code, and docs.rs didn't
have any documentation for it.

I'm not sure if this is desired yet, but I wanted to put this PR up to
solicit feedback.

Also enable doc_cfg + doc_auto_cfg so that docs.rs marks fs_utf8 (and
net) as being conditional.