Remove the description of arf-strings from fs_utf8 module comments.
arf-strings are an experiment, and not necessarily something all users
need to know about just to use fs_utf8, especially now that we have
camino. So remove the top-level comments talking about arf strings
from the module documentation.

arf-strings are still mentioned in the top-level repository README.md,
which is sufficient visibility for them for now.