README.md: Link to cap-std-ext
I am hopeful we can drain most of the functionality here into
cap-std itself or cap-fs-ext, but for now let's link to it so others
can find it more easily.

Most of this would be nice to have in Rust `std` too; that's a whole
other TODO item.  Perhaps best tackled once `std` itself depends on
rustix.