Update to io-lifetimes 0.5.1.
This updates to io-lifetimes 0.5.1, in cap-std itself and in its
dependencies.

io-lifetimes is pretty stable at this point so hopefully these
kinds of semver-breaking releases won't happen frequently. In this
case, io-lifetimes needed a version bump due to the `HandleOrNull`
change, which only affects Windows, and only affects low-level ffi
users of io-lifetimes.