tempfile: Move tests into a `#[cfg(test)] mod test`
I think having unit tests inside a module is the standard; I noticed
that this project doesn't seem to do that consistently - is there
a reason for that?

As I understand it, the main value of `#[cfg(test)] mod test {}` is
that it avoids every user of the crate pointlessly compiling the tests.