Update to io-lifetimes 0.7 and rustix 0.35.
io-lifetimes needed a semver bump due to `BorrowedFd::borrow_raw_fd`
being renamed to `BorrowedFd::borrow_raw`, to match the corresponding change
in std. cap-std doesn't use this function itself, but since it's a semver
bump, we need to bump it in cap-std so that it uses the new trait versions.

This also updates to [rustix 0.35]:

[rustix 0.35]: https://github.com/bytecodealliance/rustix/releases/tag/v0.35.6