cap-primitives: Add `read_link_contents` and `symlink_contents`
There are use cases for reading and writing the contents of
symbolic links without the check that errors out on absolute paths.
It's safe when the calling process is not going to interpret it.

For example in rpm-ostree, we construct a new root filesystem which
in some cases may have absolute links, but the calling process doesn't
read those links back.  Instead we spawn containers into the target
root to run code there.