Fix stats where `st_dev` is negative (#329)
* Fix stats where `st_dev` is negative

This is intended to be a fix for bytecodealliance/wasmtime#6978 where
some platforms use a signed integer for the `dev_t` type and it looks
like some `st_dev` fields can indeed be negative.

* Move lint location

* Comment odd location of `#[allow]`

* Fix a warning on beta Rust

* Fix a failing tempfile test

Bisection points at bytecodealliance/rustix#787 as other flags are
coming into the `Mode` enum which means that the `mode.bits()` is
`0o100644` which wasn't being tested for. The unknown bits are masked
off now.