Use custom traits instead of libstd's `MetadataExt` and `PermissionsExt`.
Define our own copy of `MetadataExt` and `PermissionsExt` and change the
code to use them instead of `std::os::*::fs::MetadataExt` and
`std::os::*::fs::PermissionsExt'.

In general, we should be moving away from using std's `Ext` traits like
this, as they weren't meant to be used in this way.

Fixes #342.