tempfile: Delegate `tempdir::new` to base version
I was looking at the code here for an unrelated reason,
and realized that there's no reason for us to duplicate the code.

Because a Dir instance doesn't have a name accessible to the calling
code, we just delegate to the non-UTF8 base version. In some presumably
very unusual situations the tempdir may have a non-UTF8 name,
but that's fine.

Signed-off-by: Colin Walters <walters@verbum.org>