[draft] Add borrowing for fs_utf8::Dir
I'd like to start using `fs_utf8::Dir`, but not in a "big bang"
fashion where I wholesale convert the entire codebase. It feels
normal and natural to have a non-utf8 `Dir` instance and
maybe sometimes just want to "view" it as utf8.

At an efficiency level, having to create a new owned fd for
this feels unnecessary.

This sketches out an implementation where we can also support
borrowing an existing fd.

This doesn't compile, but before I do a lot more typing on
this - consider it an RFC. Is this a good direction?