Don't store build check rmeta
This ensures the `.rmeta` files generated by `can_compile` are not
written to the build output. Those metadata files can create determinism
and cache invalidation issues in some build systems.

Instead the output -- since it is never actually read -- is written to
`/dev/null`.

This is a port of bytecodealliance/rustix#1200 to cap-std.